requirements.txt
streamlit>=1.38
python-dotenv>=1.0
elasticsearch>=8.12
requests>=2.31
